+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+
+## 2.3.0 - 2021-10-04
 - Use `Lmc.Logging` to create a logger
 - Allow to trace log messages into active trace
     - Add `LoggerProvider.TracingProvider` class